[feature]Support MPU6050 DMP
[what] 尝试初步适配MPU6050的DMP功能
[why]  获得更加准确的数据
[how] 添加移植的库文件

+++ User/GY_86/HMC5883L/HMC5883L.h
@@ -22,7 +22,7 @@
 
 extern int16_t HMC5883L_X,HMC5883L_Y,HMC5883L_Z;
 extern int16_t HMC5883L_X_offset,HMC5883L_Y_offset,HMC5883L_Z_offset;
-extern double HMC5883L_angle;
+extern float HMC5883L_angle;
 
 void HMC5883L_Init(void);			//模块初始化，配置相关寄存器
 void Read_HMC5883L(void);			//读取磁场数据
